requirements-lock.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements-lock.txt requirements.txt
#
affine==2.4.0
    # via
    #   rasterio
    #   xee
attrs==24.2.0
    # via rasterio
bokeh==3.5.2
    # via dask
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   netcdf4
    #   pyproj
    #   rasterio
    #   requests
cftime==1.6.4
    # via netcdf4
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   click-plugins
    #   cligj
    #   dask
    #   distributed
    #   rasterio
click-plugins==1.1.1
    # via rasterio
cligj==0.7.2
    # via rasterio
cloudpickle==3.0.0
    # via
    #   -r requirements.txt
    #   dask
    #   distributed
colorama==0.4.6
    # via click
contourpy==1.3.0
    # via bokeh
dask[array,complete,dataframe,diagnostics,distributed]==2024.8.1
    # via
    #   -r requirements.txt
    #   dask-expr
    #   distributed
    #   xarray
dask-expr==1.1.11
    # via dask
distributed==2024.8.1
    # via
    #   -r requirements.txt
    #   dask
docker==7.1.0
    # via -r requirements.txt
earthengine-api==0.1.417
    # via
    #   -r requirements.txt
    #   xee
fsspec==2024.9.0
    # via dask
google-api-core==2.20.0
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.146.0
    # via earthengine-api
google-auth==2.35.0
    # via
    #   earthengine-api
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via
    #   earthengine-api
    #   google-api-python-client
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.18.2
    # via earthengine-api
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.65.0
    # via google-api-core
httplib2==0.22.0
    # via
    #   earthengine-api
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
importlib-metadata==8.5.0
    # via dask
jinja2==3.1.4
    # via
    #   bokeh
    #   dask
    #   distributed
locket==1.0.0
    # via
    #   distributed
    #   partd
lz4==4.3.3
    # via
    #   -r requirements.txt
    #   dask
markupsafe==2.1.5
    # via jinja2
msgpack==1.1.0
    # via distributed
netcdf4==1.7.1.post2
    # via -r requirements.txt
numpy==2.0.1
    # via
    #   -r requirements.txt
    #   bokeh
    #   cftime
    #   contourpy
    #   dask
    #   netcdf4
    #   pandas
    #   pyarrow
    #   rasterio
    #   rioxarray
    #   snuggs
    #   xarray
packaging==24.1
    # via
    #   bokeh
    #   dask
    #   distributed
    #   rioxarray
    #   xarray
pandas==2.2.2
    # via
    #   -r requirements.txt
    #   bokeh
    #   dask
    #   dask-expr
    #   xarray
partd==1.4.2
    # via dask
pillow==10.4.0
    # via bokeh
proto-plus==1.24.0
    # via google-api-core
protobuf==5.28.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==6.0.0
    # via distributed
pyarrow==17.0.0
    # via
    #   dask
    #   dask-expr
pyarrow-hotfix==0.6
    # via dask
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pyparsing==3.1.4
    # via
    #   httplib2
    #   snuggs
pyproj==3.6.1
    # via
    #   rioxarray
    #   xee
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   bokeh
    #   dask
    #   distributed
rasterio==1.3.11
    # via rioxarray
requests==2.32.3
    # via
    #   docker
    #   earthengine-api
    #   google-api-core
    #   google-cloud-storage
rioxarray==0.15.0
    # via -r requirements.txt
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
snuggs==1.4.7
    # via rasterio
sortedcontainers==2.4.0
    # via distributed
tblib==3.0.0
    # via distributed
toolz==0.12.1
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.4.1
    # via
    #   -r requirements.txt
    #   bokeh
    #   distributed
tzdata==2024.1
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via
    #   distributed
    #   docker
    #   requests
xarray[parallel]==2024.9.0
    # via
    #   rioxarray
    #   xee
xee==0.0.13
    # via -r requirements.txt
xyzservices==2024.9.0
    # via bokeh
zict==3.0.0
    # via distributed
zipp==3.20.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
